requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.16
django-bootstrap-v5==1.0.11
django-pgtrigger==4.12.2
django-vite==3.0.5
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.9
PyJWT==1.7.1
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
